requirements.txt
# Основные пакеты
pandas==2.1.4
numpy==1.26.4
psycopg2-binary==2.9.9
python-dotenv==1.0.0
faker==19.13.0